requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1

numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.2.0

tensorflow==2.15.0
keras==2.15.0

psycopg2-binary==2.9.9

reportlab==4.1.0
